requirements.txt
ask-sdk-core==1.13.0
ask-sdk-model==1.22.0
ask-sdk-runtime==1.13.0
ask-sdk-s3-persistence-adapter==1.0.0
autopep8==1.5
boto3==1.12.26
botocore==1.15.26
certifi==2019.11.28
chardet==3.0.4
colorclass==2.2.0
docopt==0.6.2
docutils==0.16
idna==2.9
jmespath==0.9.5
packaging==20.3
pip-upgrader==1.4.15
pycodestyle==2.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
spotipy==2.10.0
terminaltables==3.1.0
urllib3==1.25.8
